Enforce required permissions when adding/editing known Schemes
Also move all scheme utilities into a new lib/schemeUtils.ts file
