+++ src/components/Dao/ProposalSchemeCard.tsx
@@ -4,7 +4,8 @@ import VoteGraph from "components/Proposal/Voting/VoteGraph";
 import ProposalCountdown from "components/Shared/ProposalCountdown";
 import Loading from "components/Shared/Loading";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
-import { humanProposalTitle, schemeName } from "lib/util";
+import { humanProposalTitle } from "lib/util";
+import { schemeName } from "lib/schemeUtils";
 import * as React from "react";
 import { Link } from "react-router-dom";
 import { combineLatest } from "rxjs";
