+++ src/components/Dao/SimpleSchemeCard.tsx
@@ -1,5 +1,5 @@
 import { IDAOState, ISchemeState, Scheme } from "@daostack/client";
-import { schemeName } from "lib/util";
+import { schemeName } from "lib/schemeUtils";
 import * as React from "react";
 import { Link } from "react-router-dom";
 import * as css from "./SchemeCard.scss";
