+++ src/components/Proposal/ProposalHistoryRow.tsx
@@ -3,7 +3,8 @@ import classNames from "classnames";
 import AccountPopup from "components/Account/AccountPopup";
 import AccountProfileName from "components/Account/AccountProfileName";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
-import { formatTokens, humanProposalTitle, schemeName } from "lib/util";
+import { formatTokens, humanProposalTitle } from "lib/util";
+import { schemeName } from "lib/schemeUtils";
 import * as React from "react";
 import { connect } from "react-redux";
 import { IRootState } from "reducers";
