+++ src/components/Scheme/ReputationFromToken.tsx
@@ -9,7 +9,8 @@ import { NotificationStatus } from "reducers/notifications";
 import { redeemReputationFromToken } from "actions/arcActions";
 import { enableWalletProvider, getArc } from "arc";
 import { ErrorMessage, Field, Form, Formik, FormikProps } from "formik";
-import { schemeName, fromWei } from "lib/util";
+import { fromWei } from "lib/util";
+import { schemeName } from "lib/schemeUtils";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
 import * as Sticky from "react-stickynode";
