+++ src/lib/util.ts
@@ -1,14 +1,10 @@
 import { promisify } from "util";
 import { targetedNetwork } from "arc";
-import { GenericSchemeRegistry } from "genericSchemeRegistry";
-import { rewarderContractName } from "components/Scheme/ContributionRewardExtRewarders/rewardersProps";
 import {
   Address,
-  IContractInfo,
   IProposalStage,
   IProposalState,
-  IRewardState,
-  ISchemeState} from "@daostack/client";
+  IRewardState} from "@daostack/client";
 import { of } from "rxjs";
 import { catchError } from "rxjs/operators";
 
@@ -229,114 +225,6 @@ export function sleep(milliseconds: number): Promise<void> {
   return new Promise((resolve: () => void): any => setTimeout(resolve, milliseconds));
 }
 
-/** schemes that we know how to interpret  */
-export const KNOWN_SCHEME_NAMES = [
-  "ContributionReward",
-  "GenericScheme",
-  "ReputationFromToken",
-  "SchemeRegistrar",
-  "UGenericScheme",
-  "Competition",
-  "ContributionRewardExt",
-];
-
-export const PROPOSAL_SCHEME_NAMES = [
-  "ContributionReward",
-  "GenericScheme",
-  "SchemeRegistrar",
-  "UGenericScheme",
-  "Competition",
-  "ContributionRewardExt",
-];
-
-/**
- * return true if the address is the address of a known scheme (which we know how to represent)
- * @param  address [description]
- * @return         [description]
- */
-export function isKnownScheme(address: Address) {
-  const arc = getArc();
-  let contractInfo;
-  try {
-    contractInfo = arc.getContractInfo(address);
-  } catch (err) {
-    if (err.message.match(/no contract/i)) {
-      return false;
-    }
-    throw err;
-  }
-
-  if (KNOWN_SCHEME_NAMES.includes(contractInfo.name)) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-export function schemeName(scheme: ISchemeState|IContractInfo, fallback?: string) {
-  let name: string;
-  if (scheme.name === "GenericScheme" || scheme.name === "UGenericScheme") {
-    if ((scheme as any).genericSchemeParams || ((scheme as any).uGenericSchemeParams)) {
-      const genericSchemeRegistry = new GenericSchemeRegistry();
-      let contractToCall;
-      const schemeState = scheme as ISchemeState;
-      if (schemeState.genericSchemeParams) {
-        contractToCall = schemeState.genericSchemeParams.contractToCall;
-      } else {
-        contractToCall = schemeState.uGenericSchemeParams.contractToCall;
-      }
-      const genericSchemeInfo = genericSchemeRegistry.getSchemeInfo(contractToCall);
-      if (genericSchemeInfo) {
-        name = genericSchemeInfo.specs.name;
-      } else {
-        // Adding the address is a bit long for a title
-        // name = `Blockchain Interaction (${contractToCall})`;
-        name = "Blockchain Interaction";
-      }
-    } else {
-      // this should never happen...
-      name = "Blockchain Interaction";
-    }
-  } else if (scheme.name === "ContributionReward") {
-    name ="Funding and Voting Power";
-  } else if (scheme.name === "SchemeRegistrar") {
-    name ="Plugin Manager";
-  } else if (scheme.name) {
-    if (scheme.name === "ContributionRewardExt") {
-      name = rewarderContractName(scheme as ISchemeState);
-    } else {
-      // add spaces before capital letters to approximate a human-readable title
-      name = `${scheme.name[0]}${scheme.name.slice(1).replace(/([A-Z])/g, " $1")}`;
-    }
-  } else {
-    name = fallback;
-  }
-  return name;
-}
-
-/**
- * given the address (of a scheme), return  a friendly string represeting the scheme's address and it'sname
- * @param  address [description]
- * @return         [description]
- */
-export function schemeNameAndAddress(address: string) {
-  const arc = getArc();
-  try {
-    const contractInfo = arc.getContractInfo(address);
-    const name = schemeName(contractInfo);
-
-    if (name) {
-      return `${address.slice(0, 4)}...${address.slice(-4)} (${name})`;
-    } else {
-      return `${address.slice(0, 4)}...${address.slice(-4)}`;
-    }
-  } catch (err) {
-    if (err.message.match(/No contract/)) {
-      return `${address.slice(0, 4)}...${address.slice(-4)}`;
-    }
-  }
-}
-
 /**
  * return network id, independent of the presence of Arc
  * @param web3Provider
@@ -532,68 +420,6 @@ const pattern = new RegExp(/^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w\.-]+)+[\w\-\._~:/
 
 export function isValidUrl(str: string, emptyOk = true): boolean {
   return (emptyOk && (!str || !str.trim())) || (str && pattern.test(str));
-}
-
-
-
-export enum GetSchemeIsActiveActions {
-  Register=1,
-  Remove
-}
-
-const schemeActionPropNames = new Map<string, Map<GetSchemeIsActiveActions, string>>([
-  [
-    "SchemeRegistrar" , new Map<GetSchemeIsActiveActions, string>([
-      [GetSchemeIsActiveActions.Register, "voteRegisterParams"],
-      [GetSchemeIsActiveActions.Remove, "voteRemoveParams"],
-    ]),
-  ],
-]);
-
-export function getSchemeIsActive(scheme: ISchemeState, action?: GetSchemeIsActiveActions): boolean {
-  let votingMachineParamsPropertyName: string;
-  let schemeName = `${scheme.name[0].toLowerCase()}${scheme.name.slice(1)}`;
-  if (schemeName === "genericScheme") {
-    if (scheme.uGenericSchemeParams) {
-      schemeName = "uGenericScheme";
-    }
-  }
-
-  if (action) { // then the name of the voting machine properties property depends on the action
-    const schemeActionsMap = schemeActionPropNames.get(scheme.name);
-
-    if (!schemeActionsMap) {
-      throw new Error(`getSchemeIsActive: unknown scheme: ${scheme.name}`);
-    }
-    const propName = schemeActionsMap.get(action);
-    if (!propName) {
-      throw new Error(`getSchemeIsActive: unknown action: ${scheme.name}:${action}`);
-    }
-    votingMachineParamsPropertyName = propName;
-  } else {
-    /**
-     * if scheme is SchemeRegistrar, then it is active if any of its actions are active
-     */
-    if (scheme.name === "SchemeRegistrar") {
-      return getSchemeIsActive(scheme, GetSchemeIsActiveActions.Register) || getSchemeIsActive(scheme, GetSchemeIsActiveActions.Remove);
-    } else {
-      votingMachineParamsPropertyName = "voteParams";
-    }
-  }
-
-  const schemeParams = (scheme as any)[`${schemeName}Params`][votingMachineParamsPropertyName];
-  if (!schemeParams) {
-    // eslint-disable-next-line no-console
-    console.warn(` getSchemeIsActive: scheme parameters not found at "voteParams": ${scheme.name}`);
-    return true;
-  }
-  if ((typeof(schemeParams.activationTime) === undefined) || (schemeParams.activationTime === null)) {
-    // eslint-disable-next-line no-console
-    console.warn(` getSchemeIsActive: voting machine appears not to be GenesisProtocol: ${scheme.name}`);
-    return true;
-  } else {
-    return moment(schemeParams.activationTime*1000).isSameOrBefore(moment());
-  }
 }
 
 /**
